+++ ExtensionMethods/TypeExtensions.cs
@@ -5,13 +5,18 @@ namespace KRPC.MechJeb.ExtensionMethods {
 	public static class TypeExtensions {
 		public static T CreateInstance<T>(this Type type, object[] args) {
 			try {
-				return (T)type.GetConstructor(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance, null, Type.EmptyTypes, null).Invoke(args);
+				Type[] types = Type.EmptyTypes;
+				if(args != null) {
+					types = new Type[args.Length];
+					for(int i = 0; i < args.Length; i++)
+						types[i] = args[i].GetType();
+				}
+
+				return (T)type.GetConstructor(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance, null, types, null).Invoke(args);
 			}
 			catch(Exception ex) {
 				Logger.Severe("Coudn't create an instance of " + type, ex);
-				if(ex is TargetInvocationException)
-					throw new MJServiceException(ex.Message + ": " + ex.InnerException.Message);
-				throw new MJServiceException(ex.Message);
+				throw new MJServiceException(ex.ToString());
 			}
 		}
 	}